% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targbetween.R
\name{compute_targbetween}
\alias{compute_targbetween}
\title{Compute Targeted Betweenness Centrality from Shortest Paths}
\usage{
compute_targbetween(network, paths, num_dyads, update = 0)
}
\arguments{
\item{network}{An \code{igraph} network graph}

\item{paths}{Shortest paths retrieved by
\code{\link[igraph]{all_shortest_paths}}}

\item{num_dyads}{Number of dyads used in the analysis that generated
\code{paths}}

\item{update}{A numeric value indicating how many loop iterations should
elapse between progress updates (default = \code{0}, which suppresses
output)}
}
\value{
A data.frame with the results for targeted betweenness centrality
}
\description{
This helper function takes the list of shortest paths, as retrieved by
either \code{\link{one_to_many}} or \code{\link{assorted_dyads}}, and uses it
to finish the computation of targeted betwenness centrality.
}
\details{
This function should not be called directly. Instead, provide the
\code{network} and the set of \code{dyads} that you want to analyze to
\code{\link{targbetween_alldyads}} or \code{\link{targbetween}}. Each of
those functions calls \code{\link{compute_targbetween}} as needed.
}
