% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targbetween.R
\name{one_to_many}
\alias{one_to_many}
\title{Retrieve Shortest Paths from One Source to Multiple Destinations}
\usage{
one_to_many(network, source, destination, mode)
}
\arguments{
\item{network}{An \code{igraph} network graph}

\item{source}{A single vertex or a vector of vertices in \code{network}}

\item{destination}{A vector of one or more vertices in \code{network}}

\item{mode}{A character value indicating the \code{mode} for
\code{\link[igraph]{all_shortest_paths}} (\code{"in"}, \code{"out"}, or
\code{"all"})}
}
\value{
A list of shortest paths from the \code{source} vertex to each of the
  \code{destination} vertices.
}
\description{
This helper function retrieves the set of shortest paths connecting all
possible dyads that can be formed from a single \code{source} vertex to a
vector of \code{destination} vertices, for subsequent use by
\code{\link{compute_targbetween}}.
}
\details{
This function should not be called directly. Instead, provide the network
and the set of dyads that you want to analyze to
\code{\link{targbetween_alldyads}}. That function calls
\code{\link{one_to_many}} as needed.

Note that other functions that call \code{\link{one_to_many}} may reverse
\code{source} and \code{destination} in order to identify the shortest paths
from many source vertices to a single destination vertex. When this is done
for directed networks, the \code{mode} argument is set to \code{"in"} rather
than \code{"out"} as its value. (If the network is treated as undirected in
the analysis, \code{mode = "all"} regardless of whether \code{source} and
\code{destination} are reversed.)
}
