% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targbetween.R
\name{targbetween}
\alias{targbetween}
\title{Targeted Betweenness Centrality for Specified Dyads}
\usage{
targbetween(network, dyads, filename = NULL, directed = FALSE, update = 0)
}
\arguments{
\item{network}{An \code{igraph} network graph}

\item{dyads}{A two-column data.frame, with two vertices in each row,
indicating the dyads to be considered in the analysis}

\item{filename}{A character value indicating the file location to output
results (default = \code{NULL})}

\item{directed}{A boolean value indicating whether the \code{network} should
be treated as directed (default = \code{FALSE})}

\item{update}{A numeric value indicating how many loop iterations should
elapse between progress updates (default = \code{0}, which suppresses
output)}
}
\value{
A data.frame with the results for targeted betweenness centrality
}
\description{
This function computes targeted betweenness centrality for a specified
set of dyads. \code{\link{targbetween}} is more flexible than
\code{\link{targbetween_alldyads}}, allowing different source vertices to
correspond to different destination vertices.
}
\details{
WARNING: This function calls \code{\link{assorted_dyads}}, which may be
extremely slow when the number of rows in \code{dyads} is large, as
\code{\link[igraph]{all_shortest_paths}} must be separately run for each
individual dyad. If all \code{source} vertices correspond to the same set of
\code{destination} vertices, \code{\link{targbetween_alldyads}} should be
used instead of \code{\link{targbetween}}. The
\code{\link{targbetween_alldyads}} function takes a vector of \code{source}
vertices and a vector of \code{destination} vertices as arguments, and it is
either \code{length(source)} or \code{length(destination)} times faster than
\code{\link{targbetween}}, whichever is larger.
}
\examples{
my_network <- igraph::erdos.renyi.game(20, 0.5, directed = FALSE)
various_dyads <- cbind(c(1,1,2,2,3,3,4,4,5,5),c(3,4,1,8,18,20,1,15,3,10))
results <- targbetween(my_network, various_dyads, directed=FALSE)
results
#       vertices targeted_betweenness
#  [1,] "1"      "0.2"               
#  [2,] "2"      "0"                 
#  [3,] "3"      "0.166666666666667" 
#  [4,] "4"      "0"                 
#  [5,] "5"      "0.2"               
#  [6,] "6"      "0.166666666666667" 
#  [7,] "7"      "0.166666666666667" 
#  [8,] "8"      "0.166666666666667" 
#  [9,] "9"      "0"                 
# [10,] "10"     "0.166666666666667" 
# [11,] "11"     "0.166666666666667" 
# [12,] "12"     "0.166666666666667" 
# [13,] "13"     "0.166666666666667" 
# [14,] "14"     "0.166666666666667" 
# [15,] "15"     "0.366666666666667" 
# [16,] "16"     "0.166666666666667" 
# [17,] "17"     "0"                 
# [18,] "18"     "0"                 
# [19,] "19"     "0.2"               
# [20,] "20"     "0.366666666666667" 
}
